ai-engine/requirements.txt
# Core FastAPI for AI Engine API - align with backend version
fastapi==0.136.0
uvicorn[standard]==0.44.0
python-multipart==0.0.26

# AI Framework - add upper bounds to avoid dependency conflicts
# Note: crewai requires Python 3.10-3.13. Use Docker container for development (Python 3.12).
# If running locally, use Python 3.12 virtual environment.
# Let crewai manage langchain package versions to avoid conflicts
crewai>=0.11.0,<1.0.0
openai>=1.0.0,<3.0.0
sentence-transformers>=5.1.2,<6.0.0
langchain-openai>=0.3.35  # Required by crewai for OpenAI integration

# Vector Database (pin version for embedchain compatibility)
# Note: chromadb may require Python <3.13
chromadb==1.5.8

# Data Processing - align with backend version
numpy
python-dotenv~=1.2.1

# Java analysis - tree-sitter for Java 17+ support
# Note: javalang is still required by logic_translator.py
tree-sitter>=0.23.2
tree-sitter-java
javalang==0.13.0

# File processing - align with backend version
python-magic==0.4.27
Pillow>=11.3.0
pydub
jsonschema

# Database - used by learning/correction_store and knowledge/schema
sqlalchemy>=2.0.49

# HTTP Client - align with backend version
httpx==0.28.1
beautifulsoup4>=4.14.3 # For web scraping (BedrockDocsScraper)
aiofiles>=25.1.0     # For async file operations (potentially in scraper or other utils)
validators>=0.35.0   # For URL validation (BedrockDocsScraper)

# Web Search - use available version
duckduckgo-search>=8.1.1

# Search engine
rank_bm25>=0.2.2

# Redis for job management - align with backend version
redis>=7.1.0
async-timeout>=5.0.1

# Configuration Management - align with backend version
pydantic~=2.13.3
pydantic-settings~=2.11.0

# Monitoring - align with backend version
prometheus_client>=0.25.0
structlog>=24.0.0

# Distributed Tracing - OpenTelemetry
# Note: chromadb provides opentelemetry-exporter-otlp-proto-grpc which conflicts
# with opentelemetry-exporter-jaeger-proto-grpc on protobuf versions.
# Using flexible versions and letting pip/legacy-resolver find a solution.
opentelemetry-api>=1.41.1
opentelemetry-sdk>=1.41.1
opentelemetry-exporter-otlp>=1.41.1
opentelemetry-instrumentation-fastapi>=0.62b1
opentelemetry-instrumentation-httpx>=0.62b1
opentelemetry-instrumentation-redis>=0.62b1
